zng_env

Function about

pub fn about() -> &'static About

Gets metadata about the application.

The app must call init! at the beginning of the process, otherwise the metadata will fallback to just a name extracted from the current executable file path.

See the directories::ProjectDirs::from documentation for more details on how this metadata is used to create/find the app data directories.