zng_env

Function process_name

pub fn process_name() -> Txt

Gets a process runtime name.

The primary use of this name is to identify the process in logs, see set_process_name for details about the logged name. On set or init the name is logged as an info message “pid: {pid}, name: {name}”.

Common Names

All Zng provided process handlers name the process.

"app-process" - Set by APP if no other name was set before the app starts building.

"view-process" - Set by the view-process implementer when running in multi process mode.

"crash-handler-process" - Set by the crash-handler when running with crash handling.

"crash-dialog-process" - Set by the crash-handler on the crash dialog process.

"worker-process ({worker_name}, {pid})" - Set by task worker processes if no name was set before the task runner server starts.