Function zng_env::res

pub fn res(relative_path: impl AsRef<Path>) -> PathBuf

Gets a path relative to the package resources.

The res dir can be set by init_res before any env dir is used.

In all platforms if a file bin/current_exe_name.zng_res_dir is found it defines the res path.

In cfg(debug_assertions) builds returns res.

In macOS returns bin("../Resources"), assumes the package is deployed using a desktop .app folder.

In iOS returns bin(""), assumes the package is deployed as a mobile .app folder.

In Android returns bin("../res"), assumes the package is deployed as a .apk file.

In all other Unix systems returns bin("../etc"), assumes the package is deployed using a .deb like structure.

In Windows returns bin("../res"). Note that there is no Windows standard, make sure to install the project using this structure.

Built Resources

In cfg(any(debug_assertions, feature="built_res")) builds if the target/res/{relative_path} path exists it is returned instead. This is useful during development when the app depends on res that are generated locally and not included in version control.

Note that the built resources must be packaged with the other res at the same relative location, so that release builds can find then.