zng_env

Function migrate_cache

pub fn migrate_cache(new_path: impl AsRef<Path>) -> Result<()>

Save new_path as the new cache path and make a best effort to move existing cache files.

Note that the move failure is not considered an error (it is only logged), the app is expected to rebuild missing cache entries.

Note that cache will still point to the previous path on success, the app must be restarted to use the new cache.

The new_path must not exist or be empty.